Clean-up and esm encodings
