+++ frankenmsa/align/__init__.py
@@ -1,2 +1,3 @@
 from .mmseqs_colab import MMSeqs2Colab
+from .mmseqs_local import LocalMMSeqs2Colab
 from .plm_search import PLMSearch
